Custom `EthTxManager` on `agglayer` (#109)
This PR introduces a new configuration parameter to the EthTxManager configuration on agglayer called MaxRetries. It represents a maximum number a transaction will be resent to the L1, or a maximum number of times it will try to estimate its gas until it drops it (updates the monitored transaction with the MonitoredTxStatusFailed status in the agglayer-db), meaning, it will not try to resend it again.

In order to know how many times a transaction was tried to be resent or estimated, a new column to the state.monitored_txs table needed to be added, called num_retries. That is why, a new migration file was added 0002.sql which alters given table, by adding the new column to it.